requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.1.1
django-ipware==3.0.1
django-cors-headers
django-rest-framework==0.1.0
djangorestframework==3.11.1
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.
psycopg2-binary==2.8.6
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.6.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
traitlets==5.0.4
validate-docbr==1.7.0
wcwidth==0.2.5
whitenoise==5.2.0